refactor(dashboard): promote CoverageGapBlock to common, share across tool-quality and fleet-health panels
PR #17015 introduced CoverageGapBlock inside tool-quality-panel.ts for
the rich collapsible-error UX. The fleet-health-panel Operations view
(ToolMonitorDefaultBoard) was still rendering coverage-gap with a
duplicated inline `<div>${summary}</div>${details.slice(0,2)}` — same
data, different presentation, and only the truncated first 2 details
visible.

Promote the component to `common/coverage-gap-block.ts` so both
surfaces share one implementation. Operator now sees the same warn
header / triage chips / collapsed error / RFC-link UX regardless of
which view they land on.

Changes
- New file: `dashboard/src/components/common/coverage-gap-block.ts`
  containing both `CoverageGapBlock` and `classifyCoverageError`
  (moved from tool-quality-panel.ts; pure functions colocated with
  their consumer component).
- `tool-quality-panel.ts`: -98 lines. Inline `CoverageGapBlock`
  removed, imports from common. Re-exports `classifyCoverageError`
  for test-import backward compatibility.
- `fleet-health-panel.ts`: inline 5-line render → `<CoverageGapBlock>`.
  The Operations card now shows full provenance + collapsible error,
  not just `.slice(0, 2)` of details.

The promoted module adds a `data-testid="coverage-gap-block"` for
panel-agnostic integration tests, and the classifier-hint anchor
keeps `data-testid="coverage-gap-hint-${reason}"`.

Verification
- vitest: 53/53 (tool-quality + fleet-health + source-health)
- tsc --noEmit: clean on changed files (pre-existing unrelated
  agent-roster / credential-settings errors untouched)
- eslint: 0 errors
- vite build: ✓ 11.98s

No behavior change for tool-quality view (extracted code is identical).
Fleet-health Operations view gains the richer UX as a side-effect of
the SSOT.
